Further configured the ansible workspace to include the inventory directory , groupvars, the roles directory and the playbooks directory.
We also switched to the amazon linux ami, and we configured the ansible.cfg file for passwordless ssh and host host checking so we can immediately perfrom tasks after provisioning the hosts.

Lastly we added two security groups
sg is for al three
and lb-sg is just for the load balancer.