We configured the ansible workspace again. Made sure that we used the correct key to refer tot the value of the ssh key in the ansible.cfg files Next we also made sure to reconfigure where we were saving the key, instead of inside the workspace, ust iside ~/.ssh/my_keys We also created a role (nginx) that would be called by the first playbook we are testing . The role prepares the machine and  installs nginx.
WE made sure to make the connection password less, with no host checking, become priviliges set, and just must more ease

Tommorow I tjink we sould be able to develop the role , or add one more to be used by the playbook.
The script also nly reates 3 instances, but will destroy the resources and recereate them all in the script.

We also have another script with the purpose of running the playbook.