1. Createed lambda to change dns record dynamicaly whenever a new server is provisioned and will point to its new ip 2. New roles that are more specific like install nginx, install webser, and we will cinfigure certs as soon as nginx is installed and records point to the server 3. Createed two pythonscript, one is used by the lambda, and the other, when ran, will send a post request toour api with the values of our domain, hosted zone id, and new ip address. 4. the new api will send this information to the lambda and the lambda will in turn update the records from route 53.
Next we will configure the ansible playbooks FINALLY!
ALSO used boto3 and docmunentation to create the lambda function. Very proud!

Last but ot least, we have a sample nginx conf file that we will import when everything is getting installed in the nginx web-server.

WE also need to configure ssh hardening.